backend/requirements.txt
# FastAPI Backend for Light Encounter Tabernacle

# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0

# Authentication & Security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Email
aiosmtplib==3.0.2
email-validator==2.2.0

# Environment & Configuration
pydantic-settings==2.7.1

# Utilities
python-dotenv==1.0.1
